OCT-111: moved GetAttributeOptionByCodeQuery
