+++ components/catalogs/back/src/Application/Persistence/Attribute/GetAttributeOptionsByCodeQueryInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Application\Persistence;
+namespace Akeneo\Catalogs\Application\Persistence\Attribute;
 
 /**
  * @copyright 2022 Akeneo SAS (http://www.akeneo.com)
