+++ components/catalogs/back/src/Infrastructure/Persistence/Attribute/GetAttributeOptionsByCodeQuery.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Catalogs\Infrastructure\Persistence;
+namespace Akeneo\Catalogs\Infrastructure\Persistence\Attribute;
 
-use Akeneo\Catalogs\Application\Persistence\GetAttributeOptionsByCodeQueryInterface;
+use Akeneo\Catalogs\Application\Persistence\Attribute\GetAttributeOptionsByCodeQueryInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface;
 use Akeneo\Tool\Component\StorageUtils\Repository\SearchableRepositoryInterface;
